move variables and function common to SiTrackerGSMatched and SiTrackerGSRecHit to base class, make SiTrackerGS* objects self-aware, correct initialisation of SiTrackerGSRecHits, fix sharesInput function
