+++ DataFormats/TrackerRecHit2D/src/classes_def.xml
@@ -1,17 +1,17 @@
 <lcgdict>
   <class name="GSSiTrackerRecHit2DLocalPos" ClassVersion="13">
-   <version ClassVersion="13" checksum="2851199532"/>
+   <version ClassVersion="13" checksum="641505164"/>
    <version ClassVersion="12" checksum="1712216203"/>
    <version ClassVersion="11" checksum="834943006"/>
    <version ClassVersion="10" checksum="2474445311"/>
   </class>
   <class name="SiTrackerGSRecHit2D" ClassVersion="12">
-   <version ClassVersion="12" checksum="1916407081"/>
+   <version ClassVersion="12" checksum="1304080531"/>
    <version ClassVersion="11" checksum="1613493049"/>
    <version ClassVersion="10" checksum="877294307"/>
   </class>
   <class name="SiTrackerGSMatchedRecHit2D" ClassVersion="14">
-    <version ClassVersion="14" checksum="933733669"/>
+    <version ClassVersion="14" checksum="3882739671"/>
     <version ClassVersion="13" checksum="1553184391"/>
     <version ClassVersion="12" checksum="2874379530"/>
     <version ClassVersion="11" checksum="1553184391"/>
